% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLanguageR.R
\name{cognitiveKeyPhrases}
\alias{cognitiveKeyPhrases}
\title{Key phrases.}
\usage{
cognitiveKeyPhrases(text, apiKey)
}
\arguments{
\item{text}{Text to be analyzed. Can be a dataframe or text vector.}

\item{apiKey}{API key.}
}
\value{
Extracted key phrases and/or error messages.
}
\examples{
key_phrases <- cognitiveKeyPhrases("Artificial Intelligence is regarded as one of the disruptive technologies in the 21st century", apiKey="a valid key string")
key_phrases
}
\references{
See \url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0} for more information about Text Analytics on Microsoft Cognitive Analysis.
}

