% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLanguageR.R
\name{cognitiveLangDetect}
\alias{cognitiveLangDetect}
\title{Language detection.}
\usage{
cognitiveLangDetect(text, langNum = 1, apiKey)
}
\arguments{
\item{text}{Text to be analyzed. Can be a dataframe or text vector.}

\item{langNum}{number of languages to be detected.}

\item{apiKey}{API key.}
}
\value{
Detected languages and/or error messages.
}
\examples{
lang1 <- cognitiveLangDetect("English", apiKey="a valid key string")
lang1

langs2 <- cognitiveLangDetect(c("English", "大数据"), langNum=2, apiKey="a valid key string")
langs2
}
\references{
See \url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0} for more information about Text Analytics on Microsoft Cognitive Analysis.
}

