% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLanguageR.R
\name{cognitiveSentiAnalysis}
\alias{cognitiveSentiAnalysis}
\title{Sentiment analysis.}
\usage{
cognitiveSentiAnalysis(text, apiKey)
}
\arguments{
\item{text}{Text to be analyzed. Can be a dataframe or text vector. It is recommended to process multiple document in one call so as to save cost.}

\item{apiKey}{API key. Use apiInfo("function name") to get the information.}
}
\value{
Response from sentiment analysis API, including score, text id, and/or error messages if there are.
}
\examples{
senti_score <- cognitiveSentiAnalysis("I love you", apiKey="a valid key string")
senti_score

text <- c("I love you", "I hate you", "This is love")
senti_scores <- cognitiveSentiAnalysis(text, apiKey="a valid key string")
senti_scores
}
\references{
See \url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0} for more information about Text Analytics on Microsoft Cognitive Analysis.
}

