% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLanguageR.R
\name{cognitiveTopicDetect}
\alias{cognitiveTopicDetect}
\title{Topic detection.}
\usage{
cognitiveTopicDetect(text, apiKey, stopWords = "", topicsToExclude = "")
}
\arguments{
\item{text}{Text to be analyzed. Can be a dataframe, or a text vector.}

\item{apiKey}{API key.}

\item{stopWords}{Stop words to be excluded from the input documen text.}

\item{topicsToExclude}{Topics to be excluded from the detection.}
}
\value{
Detected topics in the input documents and/or error messages.
}
\note{
Note that one transaction is charged per text document submitted. For best performance, limit each document to a short, human written text paragraph such as review, conversation or user feedback.
}
\examples{
data(text_bbc)

topics_detected <- cognitiveTopicDetect(text_bbc, apiKey="a valid key")
topics_detected
}
\references{
See \url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0} for more information about Text Analytics on Microsoft Cognitive Analysis.
}

