% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLanguageR.R
\name{cognitiveTranslation}
\alias{cognitiveTranslation}
\title{Translation.}
\usage{
cognitiveTranslation(text, lanFrom, lanTo, apiKey)
}
\arguments{
\item{text}{text to be analyzed.}

\item{lanFrom}{Source language.}

\item{lanTo}{Target language.}

\item{apiKey}{API key.}
}
\value{
A data frame containing the translated text in the target language.
}
\examples{
text <- "大数据"

translated_text1 <- cognitiveTranslation(text, lanFrom="zh-CHS", lanTo="en", apiKey="a valid key")
translated_text1

text <- "Big Data"
translated_text2 <- cognitiveTranslation(text, lanFrom="en", lanTo="zh-CHS", apiKey="a valid key")
translated_text2

translated_text3 <- cognitiveTranslation(text, lanFrom="en", lanTo="zh-CHT", apiKey="a valid key")
translated_text3
}
\references{
See \url{http://docs.microsofttranslator.com/text-translate.html#!/default/post_TranslateArray} for more information about Translation API. Get lists of languages (and codes) supported in the API at \url{https://msdn.microsoft.com/en-us/library/hh456380.aspx}.
}

